Remove warnings: m6(l/s)/m6tv(d), m6/m8(m2)/m8baby boards, drivers. 1. Remove arch m6/m6l/m6s/m6tv/m6tvd warnings. 2. Remove all m6-series boards warnings.(not include customer folder) 3. Remove parts of m8/m8m2/m8baby boards warnings.(not include customer folder) 4. Remove driver warnings.(some of usb tool by Yihui.Wu, lcd by Evoke.Zhang)

+++ arch/arm/cpu/aml_meson/m8b/lcd/lcd_config.c
@@ -996,7 +996,7 @@ static void print_lcd_clk_info(void)
     printf("vid2 pll clk      %uMHz\n"
            "lvds fifo clk     %uMHz\n"
            "cts encl clk      %uMHz\n\n",
-           clk_util_clk_msr(12), clk_util_clk_msr(24), clk_util_clk_msr(9));
+           (unsigned int)clk_util_clk_msr(12), (unsigned int)clk_util_clk_msr(24), (unsigned int)clk_util_clk_msr(9));
 }
 
 static void lcd_module_enable(void)
